Add files via upload
The goal of this study is to create QSAR models for predicting AChE inhibition activity using the ChEMBL database and machine learning algorithms. 4695 compounds were extracted from the ChEMBL database with AChE inhibition activity. We obtained various molecular descriptors and fingerprints of the compounds. The following code is for performing a mann-whitney-U test in the active compounds using the Lipinski descriptors.